% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Piece wise function}
\usage{
myf(x, coef)
}
\arguments{
\item{coef}{}
}
\value{
piece wise regression function y given x
}
\description{
Takes an x value and a coefficient and returns a piece wise regression function
}
\examples{

}
