% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Produces curve with given values}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{a}{}
}
\value{
a graph of the curve
}
\description{
Takes a mean, a sd, and an a value to show a region shaded of the curve
}
