% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Sample Package}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{time}{}
}
\value{
sample
}
\description{
Creates a sample given size, number of iterations and time between them
}
\examples{

}
